* commit '3565b8492a7081287cfbdcfcab328dd5ea3de464': (87 commits)
  Ensure integer entries are displayed in backend-import preview list (#2672)
  Show general abstain button for Y/N elections (#2675)
  Fix projection dialog projection default editing (#2662)
  Fix poll title in autoupdate (#2665)
  Update eslint rules and cleanup (#2279)
  Updated translations (#2667)
  Adjust example file topic import (#2663)
  Fix qrcode generation without password (#2657)
  Fix user creation for new assignment_candidate (#2660)
  Revert and fix meeting list account detail (#2659)
  Fix meeting list after account detail request (#2652)
  Expand cross reference links in motions (#2649)
  Add referring motions to pdf export (#2647)
  Testing action service (#2645)
  Implement WiFi access data slide (#2444)
  Fix speaker creation with newly created user (#2617)
  Fix information in user delete dialog (#2640)
  Replace collection count system (#2634)
  Created CommaSeparatedListingComponent for enumerations in html template (#2603)
  Make sort lists update only when relevant data is changed (#2472)
  ...

+++ client/src/app/site/pages/meetings/modules/list-of-speakers-content/components/list-of-speakers-content/list-of-speakers-content.component.ts
@@ -58,9 +58,9 @@ export class ListOfSpeakersContentComponent extends BaseMeetingComponent impleme
     public users: ViewUser[] = [];
     public nonAvailableUserIds: number[] = [];
 
-    public isSortMode: boolean = false;
+    public isSortMode = false;
 
-    public isMobile: boolean = false;
+    public isMobile = false;
 
     public get showFirstContributionHintObservable(): Observable<boolean> {
         return this.meetingSettingsService.get(`list_of_speakers_show_first_contribution`);
@@ -74,7 +74,7 @@ export class ListOfSpeakersContentComponent extends BaseMeetingComponent impleme
         return this.meetingSettingsService.get(`list_of_speakers_speaker_note_for_everyone`);
     }
 
-    public enableProContraSpeech: boolean = false;
+    public enableProContraSpeech = false;
 
     public get title(): string {
         return this._listOfSpeakers?.getTitle() || ``;
@@ -121,6 +121,10 @@ export class ListOfSpeakersContentComponent extends BaseMeetingComponent impleme
 
     public isCallEnabled: Observable<boolean> = this.interactionService.showLiveConfObservable;
 
+    public pointOfOrderCategoriesEnabled = false;
+
+    public restrictPointOfOrderActions = false;
+
     @Output()
     private isListOfSpeakersEmptyEvent = new EventEmitter<boolean>();
 
@@ -131,8 +135,8 @@ export class ListOfSpeakersContentComponent extends BaseMeetingComponent impleme
 
     private _listOfSpeakers: ViewListOfSpeakers | null = null;
 
-    private pointOfOrderEnabled: boolean = false;
-    private canMarkSelf: boolean = false;
+    private pointOfOrderEnabled = false;
+    private canMarkSelf = false;
 
     private get onlyPresentUsers(): boolean {
         return this.meetingSettingsService.instant(`list_of_speakers_present_users_only`) ?? false;
@@ -153,6 +157,15 @@ export class ListOfSpeakersContentComponent extends BaseMeetingComponent impleme
         private interactionService: InteractionService
     ) {
         super(componentServiceCollector, translate);
+
+        this.subscriptions.push(
+            this.meetingSettingsService
+                .get(`list_of_speakers_enable_point_of_order_categories`)
+                .subscribe(enabled => (this.pointOfOrderCategoriesEnabled = enabled)),
+            this.meetingSettingsService
+                .get(`list_of_speakers_closing_disables_point_of_order`)
+                .subscribe(enabled => (this.restrictPointOfOrderActions = enabled))
+        );
     }
 
     public ngOnInit(): void {
@@ -236,7 +249,7 @@ export class ListOfSpeakersContentComponent extends BaseMeetingComponent impleme
             if (result) {
                 await this.speakerRepo.create(this.listOfSpeakers, this._currentUser!.id, {
                     pointOfOrder: true,
-                    note: result.note
+                    ...result
                 });
             }
         } catch (e) {
@@ -247,7 +260,10 @@ export class ListOfSpeakersContentComponent extends BaseMeetingComponent impleme
     public async removePointOfOrder(): Promise<void> {
         const speakerToDelete = this.findOperatorSpeaker(true);
         if (speakerToDelete) {
-            await this.speakerRepo.delete(speakerToDelete.id);
+            const title = this.translate.instant(`Are you sure you want to irrevocably remove your point of order?`);
+            if (!(this.restrictPointOfOrderActions && this.closed) || (await this.promptService.open(title))) {
+                await this.speakerRepo.delete(speakerToDelete.id);
+            }
         }
     }
 
@@ -368,10 +384,11 @@ export class ListOfSpeakersContentComponent extends BaseMeetingComponent impleme
         const nonAvailableUsers = this.users
             .filter(
                 user =>
-                    !(!this.onlyPresentUsers || user.isPresentInMeeting()) ||
-                    this.waitingSpeakers.some(speaker => speaker.user_id === user.id)
+                    !(!this.onlyPresentUsers || user?.isPresentInMeeting()) ||
+                    this.waitingSpeakers.some(speaker => speaker.user_id === user?.id)
             )
-            .map(user => user.id);
+            .map(user => user?.id)
+            .filter(user => !!user);
         this.nonAvailableUserIds = nonAvailableUsers;
     }
 
@@ -385,7 +402,9 @@ export class ListOfSpeakersContentComponent extends BaseMeetingComponent impleme
         if (!data.userId) {
             data.userId = this.operator.operatorId;
         }
-        await this.speakerRepo.create(this.listOfSpeakers, data.userId!);
+        await this.speakerRepo.create(this.listOfSpeakers, data.userId!, {
+            meeting_user_id: data.user?.meeting_user_id
+        });
     }
 
     /**
